+++ test/integration/query/trigger-deletes.js
@@ -265,5 +265,19 @@ module.exports = function (knex) {
           });
       });
     });
+    it('should handle delete with join and trigger options', function () {
+      return knex('test_table_two')
+        .join('accounts', 'accounts.id', 'test_table_two.account_id')
+        .where({ 'accounts.email': 'test3@example.com' })
+        .del('*', triggerOptions)
+        .testSql(function (tester) {
+          tester(
+            'mssql',
+            'select top(0) [t].* into #out from [test_table_two] as t left join [test_table_two] on 0=1;delete [test_table_two] output deleted.* into #out from [test_table_two] inner join [accounts] on [accounts].[id] = [test_table_two].[account_id] where [accounts].[email] = ?; select * from #out; drop table #out;',
+            ['test3@example.com'],
+            [{ id: 3, account_id: 3, details: '', status: 1, json_data: null }]
+          );
+        });
+    });
   });
 };